Add --warning batch flag to testssl
This flag allows testssl to cancel on error.
This might be useful if a port is detected as TLS by nmap, without
actually providing an TLS/SSL enabled server.
Without this flag testssl will wait for the user to enter confirm, that
the scan should continue, even without a TLS/SSL enabled server.

Signed-off-by: Stefan Venz <stefan.venz@protonmail.com>