improve batch analysis
when no service ('-s') is specified but an output directory ('-o'),
all available services are analyzed,
using all available parsers, or the specified one ('-t').
when a service is specified but results from multiple tools are available for parsing,
either the default parser gets selected,
or all available parsers are used if an output directory is specified.

in batch mode most errors are catched and printed to STDERR and/or to the log file,
in normal mode such errors result in the termination of the analyzer.